backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
django-filter==23.4
filters==1.3.2
flake8==6.1.0
webcolors==1.11.1
djoser==2.1.0
psycopg2-binary==2.9.3
Pillow==9.0.0
pytest==6.2.4
pytest-django==4.4.0
python-dotenv
pytest-pythonpath==0.7.3
PyYAML==6.0
reportlab==4.1.0